test: add working asm.js fallback test #1
